Initial commit
Initialising an instance with a reference to the UDP ethernet object. Took a bit
of fiddling and google-digging to figure out that to initialise an object with
a member variable that is a user-defined type requires that the variable is
initialised using an initialization list. Best explanation can be found here:

http://stackoverflow.com/questions/19896259/how-to-create-class-member-objects-initialized-during-the-constructor