Make speed calculation and position update explicit
This makes the calculation of speed more explicit. Instead of the readPosition()
method causing the +current_position+ member to be updated, the position and
time are now explicitly memoized by calling update(), which also calculates the
speed, based on the time and positional difference since the last update() call.

This means that readPosition, reportPosition, readSpeed, and reportSpeed no
longer cause any recalculations and can safely be called multiple times in a
given loop, with values only being re-calculated when update() is called.